% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousSpaceDiscreteTimeModel.R
\docType{data}
\name{ContinuousSpaceDiscreteTimeModel}
\alias{ContinuousSpaceDiscreteTimeModel}
\title{Continuous space, discrete time model}
\description{
Building and estimating discrete time, continuous space models.
}
\author{
Jussi Jousimo <\email{jvj@iki.fi}>
}
\references{
Lindgren, F. & Rue, H. (2015). Bayesian Spatial Modelling with R-INLA. Journal of Statistical Software, 63(19).
}
\keyword{datasets}

