% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ContinuousSpaceModel.R
\docType{data}
\name{ContinuousSpaceModel}
\alias{ContinuousSpaceModel}
\title{Continuous space model}
\description{
Building and estimating continuous space models.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
Lindgren, F. & Rue, H. (2015). Bayesian Spatial Modelling with R-INLA. Journal of Statistical Software, 63(19).
}
\keyword{internal}

