% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kernel.R
\docType{data}
\name{ExponentialKernel}
\alias{ExponentialKernel}
\title{Exponential kernel for smoothing}
\description{
Exponential kernel for smoothing.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\keyword{datasets}

