% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NonConvexHullMesh.R
\docType{data}
\name{NonConvexHullMesh}
\alias{NonConvexHullMesh}
\title{Non-convex hull mesh}
\description{
Constructs non-convex hull mesh.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
Lindgren, F. & Rue, H. (2015). Bayesian Spatial Modelling with R-INLA. Journal of Statistical Software, 63(19).
}
\keyword{internal}

