% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpaceTimeDataFrame.R
\docType{data}
\name{SpaceTimeDataFrame}
\alias{SpaceTimeDataFrame}
\title{Space-time data frame}
\description{
Data frame for holding spatio-temporal data
}
\author{
Jussi Jousimo <\email{jvj@iki.fi}>
}
\keyword{datasets}

