% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialMesh.R
\docType{data}
\name{SpatialMesh}
\alias{SpatialMesh}
\title{2D mesh}
\description{
Constructs 2D mesh.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
Lindgren, F. & Rue, H. (2015). Bayesian Spatial Modelling with R-INLA. Journal of Statistical Software, 63(19).
}
\keyword{datasets}

