% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TwoDMesh.R
\docType{data}
\name{TwoDMesh}
\alias{TwoDMesh}
\title{2D mesh}
\description{
Constructs 2D mesh.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
Lindgren, F. & Rue, H. (2015). Bayesian Spatial Modelling with R-INLA. Journal of Statistical Software, 63(19).
}
\keyword{internal}

